pyproject.toml
[tool.poetry]
name = "kinomir"
version = "0.1.0"
description = ""
authors = ["Denis Tokarev <wf.denis.tokarev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
django-debug-toolbar = "^4.4.6"
gunicorn = "^23.0.0"
django-rest-framework = "^0.1.0"
mypy = "^1.14.0"
django-stubs = "^5.1.1"
djangorestframework-stubs = "^3.15.2"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
